Fix bug on throwing the ball to eliminated player
Originally, the code used the thrower and target id's as direct inputs
to the get_child method. However, that stopped working correctly once a
player in the middle of the list got eliminated and removed from it.

To fix this, I then used the id's to get the correct player nodes via
their name (which always starts with "Player" and ends with their ID).
However, this still didn't work correctly when one of those nodes was
already freed when the callback was run, e.g. because the target was
eliminated by an enemy. In those instances the ball woulde freeze and
the game effectively bricke and could only be restarted.

To prevent this, this new version now uses get_node_or_null instead and
returns if either the thrower or the target end up as null.

Also, removes some (unused) lines that would introduce one player after
the other by making their controller vibrate for a second. On repeated
restarts this 2-4 second delay (depending on the player count) quickly
became tedious and annoying.